+++ CHANGELOG.md
@@ -2,6 +2,35 @@
 
 <!--next-version-placeholder-->
 
+## v0.3.0 (2023-01-29)
+### Feature
+* **twitter-persona:** Prepare release for milestone 3 ([`c352228`](https://github.com/UBC-MDS/twitter-persona/commit/c352228cd6ee8cb0741cb878fbdc87282ff34439))
+
+## v0.2.4 (2023-01-29)
+
+
+## v0.2.3 (2023-01-29)
+
+
+## v0.2.2 (2023-01-28)
+### Fix
+* **preprocessing doc:** Added docs for tests funs ([`d2b9a74`](https://github.com/UBC-MDS/twitter-persona/commit/d2b9a74a2f1193f5ae9be4b2c1f87ae2f7838fcd))
+
+## v0.2.1 (2023-01-27)
+
+
+## v0.2.0 (2023-01-27)
+### Feature
+* **package:** Update code in load_twitter_msg ([`1045592`](https://github.com/UBC-MDS/twitter-persona/commit/10455929f03f2e15370173f5598ff49e28239cf8))
+
+## v0.1.0 (2023-01-27)
+### Feature
+* **package:** Add code in load_twitter_msg ([`77face9`](https://github.com/UBC-MDS/twitter-persona/commit/77face9241dc89bdca242f2a513b9b594f34c874))
+* **package:** Add new function test ([`dc7721f`](https://github.com/UBC-MDS/twitter-persona/commit/dc7721f7bc06ba418713299269bcfef58a3114ab))
+
+## v0.0.1 (2023-01-27)
+
+
 ## v0.1.0 (11/01/2023)
 
 - First release of `twitterpersona`!